feat(LevelState): paused text
For this to work I had to make the cursor and the backdrop components, and alter the code to adapt to that. Refactoring needed.

Also making the cursor a component made it easier (but sometimes a bit more confusing) to replicate the fake cursor functionality of LevelState, which is now also present in MenuState.