feat(Sprites): pipes
For this to be added, I had to remove the more variants in the remastered fidelity, in order to stop adding more inconsistencies, since I would have to make new sprites, and it would make pipes and empty blocks have variants, but not the rest of the blocks. At least for now it will remain this way, but I may add the variants later on.